Landmark Recognition using DELF.
For every image, the closest estimates of landmark are made using imagehash. The nearest landmarks are found by assigning weight based on hash difference. DELF comparison is made for all the landmarks with the image and the sum of inliers is chosen as the criteria. Thresholds are set to classify as one of the shortlisted landmarks or not a landmark.